+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Gov.hs
@@ -51,8 +51,6 @@ import Cardano.Ledger.BaseTypes (
 import Cardano.Ledger.Binary (
   DecCBOR (..),
   EncCBOR (..),
-  FromCBOR (..),
-  ToCBOR (..),
   internMap,
   internSet,
  )
@@ -301,12 +299,6 @@ instance EraPParams era => EncCBOR (ConwayGovPredFailure era) where
         Sum TreasuryWithdrawalReturnAccountsDoNotExist 17 !> To accounts
       UnelectedCommitteeVoters committee ->
         Sum UnelectedCommitteeVoters 18 !> To committee
-
-instance EraPParams era => ToCBOR (ConwayGovPredFailure era) where
-  toCBOR = toEraCBOR @era
-
-instance EraPParams era => FromCBOR (ConwayGovPredFailure era) where
-  fromCBOR = fromEraCBOR @era
 
 data ConwayGovEvent era
   = GovNewProposals !TxId !(Proposals era)