+++ libs/cardano-ledger-core/src/Cardano/Ledger/Keys/WitVKey.hs
@@ -15,14 +15,14 @@ module Cardano.Ledger.Keys.WitVKey (
 
 import Cardano.Crypto.DSIGN.Class (
   SignedDSIGN,
-  encodeSignedDSIGN,
  )
 import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR (..),
   EncCBOR (..),
+  encodeListLen,
   shelleyProtVer,
-  toPlainDecoder,
+  toPlainEncoding,
  )
 import Cardano.Ledger.Binary.Coders (Decode (..), decode, (<!))
 import qualified Cardano.Ledger.Binary.Plain as Plain
@@ -72,16 +72,13 @@ instance Typeable kr => Ord (WitVKey kr) where
     comparing wvkKeyHash x y <> comparing (hashTxBodySignature . wvkSignature) x y
 
 instance Typeable kr => Plain.ToCBOR (WitVKey kr) where
-  toCBOR WitVKeyInternal {wvkKey, wvkSignature} =
-    Plain.encodeListLen 2
-      <> Plain.toCBOR wvkKey
-      <> encodeSignedDSIGN wvkSignature
+  toCBOR = toPlainEncoding shelleyProtVer . encCBOR
 
-instance Typeable kr => Plain.FromCBOR (WitVKey kr) where
-  fromCBOR = toPlainDecoder Nothing shelleyProtVer decCBOR
-  {-# INLINE fromCBOR #-}
-
-instance Typeable kr => EncCBOR (WitVKey kr)
+instance EncCBOR (WitVKey kr) where
+  encCBOR (WitVKey k sig) =
+    encodeListLen 2
+      <> encCBOR k
+      <> encCBOR sig
 
 instance Typeable kr => DecCBOR (WitVKey kr) where
   decCBOR = decode (RecD WitVKey <! From <! From)